+++ aisdc/attacks/likelihood_attack.py
@@ -1,16 +1,14 @@
 """Likelihood testing scenario from https://arxiv.org/pdf/2112.03570.pdf."""
 
-# pylint: disable = too-many-branches
-
 from __future__ import annotations
 
+import contextlib
 import logging
-from collections.abc import Iterable
 
 import numpy as np
 import sklearn
 from fpdf import FPDF
-from scipy.stats import norm
+from scipy.stats import norm, shapiro
 
 from aisdc import metrics
 from aisdc.attacks import report
@@ -20,55 +18,21 @@ from aisdc.attacks.target import Target
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-EPS = 1e-16  # Used to avoid numerical issues in logit function
-
-
-class DummyClassifier:
-    """A Dummy Classifier to allow this code to work with get_metrics."""
-
-    def predict(self, X_test):
-        """Return an array of 1/0 depending on value in second column."""
-        return 1 * (X_test[:, 1] > 0.5)
-
-    def predict_proba(self, X_test):
-        """Simply return the X_test."""
-        return X_test
-
-
-def _logit(p: float) -> float:
-    """Return standard logit.
-
-    Parameters
-    ----------
-    p : float
-        value to evaluate logit at
-
-    Returns
-    -------
-    li : float
-        logit(p)
-
-    Notes
-    -----
-    If p is close to 0 or 1, evaluating the log will result in numerical
-    instabilities.  This code thresholds p at EPS and 1 - EPS where EPS
-    defaults at 1e-16.
-    """
-    if p > 1 - EPS:  # pylint:disable=consider-using-min-builtin
-        p = 1 - EPS
-    p = max(p, EPS)
-    return np.log(p / (1 - p))
+EPS = 1e-16  # Used to avoid numerical issues
 
 
 class LIRAAttack(Attack):
     """The main LiRA Attack class."""
 
-    def __init__(
+    def __init__(  # pylint: disable=too-many-arguments
         self,
         output_dir: str = "outputs",
         write_report: bool = True,
         n_shadow_models: int = 100,
         p_thresh: float = 0.05,
+        mode: str = "offline",
+        fix_variance: bool = False,
+        report_individual: bool = False,
     ) -> None:
         """Construct an object to execute a LiRA attack.
 
@@ -83,10 +47,33 @@ class LIRAAttack(Attack):
         p_thresh : float
             Threshold to determine significance of things. For instance
             auc_p_value and pdif_vals.
+        mode : str
+            Attack mode: {"offline", "offline-carlini", "online-carlini"}
+        fix_variance : bool
+            Whether to use the global standard deviation or per record.
+        report_individual : bool
+            Whether to report metrics for each individual record.
         """
         super().__init__(output_dir=output_dir, write_report=write_report)
-        self.n_shadow_models = n_shadow_models
-        self.p_thresh = p_thresh
+        self.n_shadow_models: int = n_shadow_models
+        self.p_thresh: float = p_thresh
+        self.mode: str = mode
+        self.fix_variance: bool = fix_variance
+        self.report_individual: bool = report_individual
+
+        self.result: dict = {}  # individual record results
+        if self.report_individual:
+            self.result["score"] = []
+            self.result["label"] = []
+            self.result["target_logit"] = []
+            self.result["out_p_norm"] = []
+            self.result["out_prob"] = []
+            self.result["out_mean"] = []
+            self.result["out_std"] = []
+            if self.mode == "online-carlini":
+                self.result["in_prob"] = []
+                self.result["in_mean"] = []
+                self.result["in_std"] = []
 
     def __str__(self):
         """Return the name of the attack."""
@@ -94,8 +81,6 @@ class LIRAAttack(Attack):
 
     def attack(self, target: Target) -> dict:
         """Run a LiRA attack from a Target object and a target model.
-
-        Needs to have X_train, X_test, y_train and y_test set.
 
         Parameters
         ----------
@@ -107,10 +92,11 @@ class LIRAAttack(Attack):
         dict
             Attack report.
         """
+        # prepare
         shadow_clf = sklearn.base.clone(target.model)
         target = self._check_and_update_dataset(target)
         # execute attack
-        self.run_scenario_from_preds(
+        self._run(
             shadow_clf,
             target.X_train,
             target.y_train,
@@ -139,12 +125,10 @@ class LIRAAttack(Attack):
         classes = list(target.model.classes_)
         for y in target.y_train:
             y_train_new.append(classes.index(y))
-
         target.y_train = np.array(y_train_new, int)
-
         logger.info(
-            "new ytrain has values and counts: %s",
-            f"{np.unique(target.y_train,return_counts=True)}",
+            "new y_train has values and counts: %s",
+            np.unique(target.y_train, return_counts=True),
         )
         ok_pos = []
         y_test_new = []
@@ -152,144 +136,247 @@ class LIRAAttack(Attack):
             if y in classes:
                 ok_pos.append(i)
                 y_test_new.append(classes.index(y))
-
         if len(y_test_new) != len(target.X_test):
             target.X_test = target.X_test[ok_pos, :]
         target.y_test = np.array(y_test_new, int)
         logger.info(
-            "new ytest has values and counts: %s",
-            f"{np.unique(target.y_test,return_counts=True)}",
+            "new y_test has values and counts: %s",
+            np.unique(target.y_test, return_counts=True),
         )
-
         return target
 
-    def run_scenario_from_preds(  # pylint: disable = too-many-statements, too-many-arguments, too-many-locals
+    def _run(  # pylint: disable=too-many-arguments,too-many-locals
         self,
         shadow_clf: sklearn.base.BaseEstimator,
-        X_target_train: Iterable[float],
-        y_target_train: Iterable[float],
-        target_train_preds: Iterable[float],
-        X_shadow_train: Iterable[float],
-        y_shadow_train: Iterable[float],
-        shadow_train_preds: Iterable[float],
+        X_train: np.ndarray,
+        y_train: np.ndarray,
+        proba_train: np.ndarray,
+        X_test: np.ndarray,
+        y_test: np.ndarray,
+        proba_test: np.ndarray,
     ) -> None:
-        """Run the likelihood test, using the "offline" version.
+        """Run the likelihood test.
 
         See p.6 (top of second column) for details.
+
+        With mode "offline", we measure the probability of observing a
+        confidence as high as the target model's under the null-hypothesis that
+        the target point is a non-member. That is we, use the norm CDF.
+
+        With mode "offline-carlini", we measure the probability that a target point
+        did not come from the non-member distribution. That is, we use Carlini's
+        implementation with a single norm (log) PDF.
+
+        With mode "online-carlini", we use Carlini's implementation of the standard
+        likelihood ratio test, measuring the ratio of probabilities the sample came
+        from the two distributions. That is, the (log) PDF of pr_in minus pr_out.
 
         Parameters
         ----------
         shadow_clf : sklearn.Model
-            An sklearn classifier that will be trained to form the shadow model.
-            All hyper-parameters should have been set.
-        X_target_train : np.ndarray
-            Data that was used to train the target model
-        y_target_train : np.ndarray
-            Labels that were used to train the target model
-        target_train_preds : np.ndarray
-            Array of predictions produced by the target model on the training data
-        X_shadow_train : np.ndarray
-            Data that will be used to train the shadow models
-        y_shadow_train : np.ndarray
-            Labels that will be used to train the shadow model
-        shadow_train_preds : np.ndarray
-            Array of predictions produced by the target model on the shadow data
+            An sklearn classifier that will be trained to form the shadow models.
+            All hyperparameters should have been set.
+        X_train : np.ndarray
+            Data that was used to train the target model.
+        y_train : np.ndarray
+            Labels that were used to train the target model.
+        proba_train : np.ndarray
+            Array of predictions produced by the target model on the training data.
+        X_test : np.ndarray
+            Data that will be used to train the shadow models.
+        y_test : np.ndarray
+            Labels that will be used to train the shadow models.
+        proba_test : np.ndarray
+            Array of predictions produced by the target model on the shadow data.
         """
-        n_train_rows, _ = X_target_train.shape
-        n_shadow_rows, _ = X_shadow_train.shape
-        indices = np.arange(0, n_train_rows + n_shadow_rows, 1)
+        logger.info("Running %s LiRA, fix_variance=%s", self.mode, self.fix_variance)
 
-        # Combine taregt and shadow train, from which to sample datasets
-        combined_x_train = np.vstack((X_target_train, X_shadow_train))
-        combined_y_train = np.hstack((y_target_train, y_shadow_train))
+        # Combine target and shadow train, from which to sample datasets
+        n_train_rows, _ = X_train.shape
+        n_shadow_rows, _ = X_test.shape
+        combined_x_train = np.vstack((X_train, X_test))
+        combined_y_train = np.hstack((y_train, y_test))
+        combined_target_preds = np.vstack((proba_train, proba_test))
 
-        train_row_to_confidence = {i: [] for i in range(n_train_rows)}
-        shadow_row_to_confidence = {i: [] for i in range(n_shadow_rows)}
+        # Get the confidences of samples when in and not in the training set
+        out_conf, in_conf = self._train_shadow_models(
+            shadow_clf,
+            combined_x_train,
+            combined_y_train,
+            n_train_rows,
+        )
 
-        # Train N_SHADOW_MODELS shadow models
+        # Get the LiRA scores, and how many confidences were normally distributed
+        mia_scores, n_normal = self._compute_scores(
+            combined_y_train, combined_target_preds, out_conf, in_conf
+        )
+
+        # Save metrics
+        mia_clf = self._DummyClassifier()
+        mia_scores = np.array(mia_scores)
+        mia_labels = np.array([1] * n_train_rows + [0] * n_shadow_rows)
+        y_pred_proba = mia_clf.predict_proba(mia_scores)
+        self.attack_metrics = [metrics.get_metrics(y_pred_proba, mia_labels)]
+        self.attack_metrics[-1]["n_normal"] = n_normal / (n_train_rows + n_shadow_rows)
+        if self.report_individual:
+            self.result["score"] = [score[1] for score in mia_scores]
+            self.result["member"] = mia_labels
+            self.attack_metrics[-1]["individual"] = self.result
+
+        logger.info("Finished scenario")
+
+    def _compute_scores(  # pylint: disable=too-many-locals
+        self,
+        combined_y_train: np.ndarray,
+        combined_target_preds: np.ndarray,
+        out_conf: dict[list[float]],
+        in_conf: dict[list[float]],
+    ) -> tuple[list[list[float]], int]:
+        """Compute LiRA scores for each record."""
+        logger.info("Computing scores")
+
+        mia_scores: list[list[float]] = []
+        n_normal: int = 0
+        global_in_std: float = self._get_global_std(in_conf)
+        global_out_std: float = self._get_global_std(out_conf)
+
+        # score each record in the member and non-member sets
+        for i, label in enumerate(combined_y_train):
+            # get the target model behaviour on the record
+            target_logit: float = _logit(combined_target_preds[i, label])
+            # get behaviour observed with the record as a non-member
+            out_mean, out_std = self._describe_conf(out_conf[i], global_out_std)
+            # get behaviour observed with the record as a member
+            in_mean, in_std = self._describe_conf(in_conf[i], global_in_std)
+            # compare behaviour
+            if self.mode == "offline":
+                pr_out = norm.cdf(target_logit, loc=out_mean, scale=out_std + EPS)
+                pr_in = 1 - pr_out
+            elif self.mode == "online-carlini":
+                pr_out = -norm.logpdf(target_logit, out_mean, out_std + EPS)
+                pr_in = -norm.logpdf(target_logit, in_mean, in_std + EPS)
+                # ratio
+                pr_in = pr_in - pr_out
+                pr_out = -pr_in
+            elif self.mode == "offline-carlini":
+                pr_out = -norm.logpdf(target_logit, out_mean, out_std + EPS)
+                pr_in = -pr_out
+            else:
+                raise ValueError(f"Unsupported LiRA mode: {self.mode}")
+            mia_scores.append([pr_in, pr_out])
+            # test the non-member samples for normality
+            out_p_norm = self._get_p_normal(np.array(out_conf[i]))
+            if out_p_norm <= 0.05:
+                n_normal += 1
+            # save individual record result
+            if self.report_individual:
+                self.result["label"].append(label)
+                self.result["target_logit"].append(target_logit)
+                self.result["out_p_norm"].append(out_p_norm)
+                self.result["out_prob"].append(pr_out)
+                self.result["out_mean"].append(out_mean)
+                self.result["out_std"].append(out_std + EPS)
+                if self.mode == "online-carlini":
+                    self.result["in_prob"].append(pr_in)
+                    self.result["in_mean"].append(in_mean)
+                    self.result["in_std"].append(in_std + EPS)
+        return mia_scores, n_normal
+
+    def _describe_conf(
+        self, confidences: list[float], global_std: float
+    ) -> tuple[float, float]:
+        """Return the mean and standard deviation of a list of confidences."""
+        scores: np.ndarray = np.array(confidences)
+        mean: float = 0
+        std: float = 0
+        if not np.isnan(scores).all():
+            mean = np.nanmean(scores)
+            std = np.nanstd(scores)
+        if self.fix_variance:
+            std = global_std
+        return mean, std
+
+    def _get_global_std(self, confidences: dict[str, list[float]]) -> float:
+        """Return the global standard deviation."""
+        global_std: float = 0
+        if self.fix_variance:
+            # requires conversion from a dict of diff size proba lists
+            arrays = list(confidences.values())
+            combined = np.concatenate(arrays)
+            if not np.isnan(combined).all():
+                global_std = np.nanstd(combined)
+        return global_std
+
+    def _get_p_normal(self, samples: np.ndarray) -> float:
+        """Test whether a set of samples is normally distributed."""
+        p_normal: float = np.NaN
+        if np.nanvar(samples) > EPS:
+            with contextlib.suppress(ValueError):
+                _, p_normal = shapiro(samples)
+        return p_normal
+
+    def _train_shadow_models(  # pylint: disable=too-many-locals
+        self,
+        shadow_clf: sklearn.base.BaseEstimator,
+        combined_x_train: np.ndarray,
+        combined_y_train: np.ndarray,
+        n_train_rows: int,
+    ) -> tuple[dict, dict]:
+        """Train shadow models and return confidence scores.
+
+        Parameters
+        ----------
+        shadow_clf : sklearn.base.BaseEstimator
+            An sklearn classifier that will be trained to form the shadow models.
+        combined_x_train : np.ndarray
+            Array of combined train and test features.
+        combined_y_train : np.ndarray
+            Array of combined train and test labels.
+        n_train_rows : int
+            Number of samples in the training set.
+
+        Returns
+        -------
+        tuple[dict, dict]
+            Dictionary of confidences when not in the training set.
+            Dictionary of confidences when in the training set.
+        """
         logger.info("Training shadow models")
+
+        n_combined, _ = combined_x_train.shape
+        out_conf: dict = {i: [] for i in range(n_combined)}
+        in_conf: dict = {i: [] for i in range(n_combined)}
+        indices: np.ndarray = np.arange(0, n_combined, 1)
+
         for model_idx in range(self.n_shadow_models):
             if model_idx % 10 == 0:
                 logger.info("Trained %d models", model_idx)
             # Pick the indices to use for training this one
             np.random.seed(model_idx)  # Reproducibility
             these_idx = np.random.choice(indices, n_train_rows, replace=False)
-            temp_x_train = combined_x_train[these_idx, :]
-            temp_y_train = combined_y_train[these_idx]
-
             # Fit the shadow model
             shadow_clf.set_params(random_state=model_idx)
-            shadow_clf.fit(temp_x_train, temp_y_train)
-
+            shadow_clf.fit(
+                combined_x_train[these_idx, :],
+                combined_y_train[these_idx],
+            )
             # map a class to a column
             class_map = {c: i for i, c in enumerate(shadow_clf.classes_)}
-
-            # Get the predicted probabilities on the training data
-            confidences = shadow_clf.predict_proba(X_target_train)
-
+            # generate shadow confidences
+            shadow_confidences = shadow_clf.predict_proba(combined_x_train)
             these_idx = set(these_idx)
-            for i in range(n_train_rows):
+            for i, conf in enumerate(shadow_confidences):
+                # logit of the correct class
+                label = class_map.get(combined_y_train[i], -1)
+                # Occasionally, the random data split will result in classes being
+                # absent from the training set. In these cases label will be -1 and
+                # we include logit(0) instead of discarding
+                logit = _logit(0) if label < 0 else _logit(conf[label])
                 if i not in these_idx:
-                    # If i was _not_ used for training, incorporate the logit of its confidence of
-                    # being correct - TODO: should we just be taking max??
-                    cl_pos = class_map.get(y_target_train[i], -1)
-                    # Occasionally, the random data split will result in classes being
-                    # absent from the training set. In these cases cl_pos will be -1 and
-                    # we include logit(0) instead of discarding (also for the shadow data below)
-                    if cl_pos >= 0:
-                        train_row_to_confidence[i].append(
-                            _logit(confidences[i, cl_pos])
-                        )
-                    else:  # pragma: no cover
-                        # catch-all
-                        train_row_to_confidence[i].append(_logit(0))
-            # Same process for shadow data
-            shadow_confidences = shadow_clf.predict_proba(X_shadow_train)
-            for i in range(n_shadow_rows):
-                if i + n_train_rows not in these_idx:
-                    cl_pos = class_map.get(y_shadow_train[i], -1)
-                    if cl_pos >= 0:
-                        shadow_row_to_confidence[i].append(
-                            _logit(shadow_confidences[i, cl_pos])
-                        )
-                    else:  # pragma: no cover
-                        # catch-all
-                        shadow_row_to_confidence[i].append(_logit(0))
-
-        # Do the test described in the paper in each case
-        mia_scores = []
-        mia_labels = []
-        logger.info("Computing scores for train rows")
-        for i in range(n_train_rows):
-            true_score = _logit(target_train_preds[i, y_target_train[i]])
-            null_scores = np.array(train_row_to_confidence[i])
-            mean_null = 0.0
-            var_null = 0.0
-            if not np.isnan(null_scores).all():
-                mean_null = np.nanmean(null_scores)  # null_scores.mean()
-                var_null = np.nanvar(null_scores)  #
-            var_null = max(var_null, EPS)  # var can be zero in some cases
-            prob = norm.cdf(true_score, loc=mean_null, scale=np.sqrt(var_null))
-            mia_scores.append([1 - prob, prob])
-            mia_labels.append(1)
-
-        logger.info("Computing scores for shadow rows")
-        for i in range(n_shadow_rows):
-            true_score = _logit(shadow_train_preds[i, y_shadow_train[i]])
-            null_scores = np.array(shadow_row_to_confidence[i])
-            mean_null = null_scores.mean()
-            var_null = max(null_scores.var(), EPS)  # var can be zeros in some cases
-            prob = norm.cdf(true_score, loc=mean_null, scale=np.sqrt(var_null))
-            mia_scores.append([1 - prob, prob])
-            mia_labels.append(0)
-
-        mia_clf = DummyClassifier()
-        logger.info("Finished scenario")
-
-        mia_scores = np.array(mia_scores)
-        mia_labels = np.array(mia_labels)
-        y_pred_proba = mia_clf.predict_proba(mia_scores)
-        self.attack_metrics = [metrics.get_metrics(y_pred_proba, mia_labels)]
+                    out_conf[i].append(logit)
+                else:
+                    in_conf[i].append(logit)
+        return out_conf, in_conf
 
     def _construct_metadata(self) -> None:
         """Construct the metadata object."""
@@ -325,9 +412,43 @@ class LIRAAttack(Attack):
         """Construct the metadata object after attacks."""
         attack_metrics_experiment = {}
         attack_metrics_instances = {}
-
         for rep, _ in enumerate(self.attack_metrics):
             attack_metrics_instances["instance_" + str(rep)] = self.attack_metrics[rep]
-
         attack_metrics_experiment["attack_instance_logger"] = attack_metrics_instances
         return attack_metrics_experiment
+
+    class _DummyClassifier:
+        """A Dummy Classifier to allow this code to work with get_metrics."""
+
+        def predict(self, X_test):
+            """Return an array of 1/0 depending on value in second column."""
+            return 1 * (X_test[:, 1] > 0.5)
+
+        def predict_proba(self, X_test):
+            """Simply return the X_test."""
+            return X_test
+
+
+def _logit(p: float) -> float:
+    """Return standard logit.
+
+    Parameters
+    ----------
+    p : float
+        value to evaluate logit at.
+
+    Returns
+    -------
+    float
+        logit(p)
+
+    Notes
+    -----
+    If `p` is close to 0 or 1, evaluating the log will result in numerical
+    instabilities. This code thresholds `p` at `EPS` and `1 - EPS` where `EPS`
+    defaults at 1e-16.
+    """
+    if p > 1 - EPS:
+        p = 1 - EPS
+    p = max(p, EPS)
+    return np.log(p / (1 - p))